reorganise deleteResource
