+++ packages/containers/src/DeleteResource/DeleteResource.connect.js
