+++ packages/containers/src/DeleteResource/DeleteResource.container.js
