+++ packages/containers/src/DeleteResource/deleteResource.constants.js
