+++ packages/containers/src/DeleteResource/index.js
